move process_events to chain subscriber
